move AddTag to pages
